# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A social media API service built with Go featuring user management, posts, comments, and following functionality",
        "title": "Gopher Social API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/authentication/token": {
            "post": {
                "description": "Authenticate user with email and password, returns a JWT token for API access",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Login and get authentication token",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_payload.CreateUserTokenPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "JWT token created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_payload.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/authentication/user": {
            "post": {
                "description": "Create a new user account with username, email and password. Returns user information with an activation token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_payload.RegisterUserPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully, activation required",
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_payload.UserWithToken"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error or duplicate email/username",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_orangeMangoDimz_go-social_internal_entities_payload.CreateUserTokenPayload": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email address",
                    "type": "string",
                    "maxLength": 255,
                    "example": "johndoe@example.com"
                },
                "password": {
                    "description": "Password",
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 3,
                    "example": "securepassword123"
                }
            }
        },
        "github_com_orangeMangoDimz_go-social_internal_entities_payload.RegisterUserPayload": {
            "description": "Request payload for user registration",
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "Email address (max 255 characters)",
                    "type": "string",
                    "maxLength": 255,
                    "example": "johndoe@example.com"
                },
                "password": {
                    "description": "Password (3-72 characters)",
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 3,
                    "example": "securepassword123"
                },
                "username": {
                    "description": "Username (max 100 characters)",
                    "type": "string",
                    "maxLength": 100,
                    "example": "johndoe"
                }
            }
        },
        "github_com_orangeMangoDimz_go-social_internal_entities_payload.TokenResponse": {
            "description": "Response containing a JWT authentication token",
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT authentication token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
        "github_com_orangeMangoDimz_go-social_internal_entities_payload.UserWithToken": {
            "description": "User information with activation token",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Account creation timestamp",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "email": {
                    "description": "Email address",
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "description": "User ID",
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_users.Role"
                },
                "role_id": {
                    "type": "integer"
                },
                "token": {
                    "description": "Activation token",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "github_com_orangeMangoDimz_go-social_internal_entities_users.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for authorization",
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Enter your bearer token in the format **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}